move `create_agg_state_manager` to `AggStateManager::create`
